+++ src/main/java/application/game/Board.java
@@ -1,5 +1,6 @@
-package application;
+package application.game;
 
+import application.ImmutablePosition;
 import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableList;
 import org.springframework.beans.factory.annotation.Autowired;
