+++ src/main/java/application/game/COLOUR.java
@@ -1,10 +1,10 @@
-package application;
+package application.game;
 
 public enum COLOUR {
     BLACK,
     WHITE;
 
-    static COLOUR opposite(COLOUR colour) {
+    public static COLOUR opposite(COLOUR colour) {
         if (colour.equals(WHITE)) {
             return BLACK;
         } else {
