+++ src/main/java/application/game/Position.java
@@ -1,4 +1,4 @@
-package application;
+package application.game;
 
 import org.immutables.value.Value;
