+++ src/main/java/application/players/ComputerPlayer.java
@@ -1,5 +1,12 @@
-package application;
+package application.players;
 
+import application.ImmutablePosition;
+import application.game.Board;
+import application.game.COLOUR;
+import application.game.Counter;
+import application.mcts.MonteCarloTreeSearch;
+import application.mcts.TreeNode;
+import application.utils.MessageProducer;
 import com.google.common.collect.ImmutableList;
 
 import java.util.Random;
@@ -27,7 +34,7 @@ public class ComputerPlayer implements Player {
         System.out.println(counterColour + "'s Turn");
         ImmutablePosition position;
         if (this.moveFunction.equals(1)) {
-            position = getNextPositionHuerstic(board);
+            position = getNextPositionHeuristic(board);
         } else if (this.moveFunction.equals(2)) {
             position = getNextPositionMCTS(board);
         } else {
@@ -47,7 +54,7 @@ public class ComputerPlayer implements Player {
     public void playTurnKafka(Board board) {
         boolean invalidMove = true;
         while (invalidMove) {
-            ImmutablePosition position = getNextPositionHuerstic(board);
+            ImmutablePosition position = getNextPositionHeuristic(board);
             Counter counter = new Counter(counterColour);
             if (board.addCounter(counter, position)) {
                 invalidMove = false;
@@ -67,17 +74,17 @@ public class ComputerPlayer implements Player {
         return counterColour;
     }
 
-    private ImmutablePosition getNextPositionHuerstic(Board board) {
+    private ImmutablePosition getNextPositionHeuristic(Board board) {
         ImmutableList<ImmutablePosition> validMoves = board.getValidMoves(this.counterColour);
         Counter counter = new Counter(this.counterColour);
-        Double bestBoardHeurstic = Double.MIN_VALUE;
+        Double bestBoardHeuristic = Double.MIN_VALUE;
         ImmutablePosition bestMove = null;
         for (ImmutablePosition position : validMoves) {
             Board futureBoard = board.clone();
             futureBoard.addCounter(counter, position);
-            Double boardHeurstic = futureBoard.getBoardHeuristic(this.counterColour, 1);
-            if (boardHeurstic > bestBoardHeurstic) {
-                bestBoardHeurstic = boardHeurstic;
+            Double boardHeuristic = futureBoard.getBoardHeuristic(this.counterColour, 1);
+            if (boardHeuristic > bestBoardHeuristic) {
+                bestBoardHeuristic = boardHeuristic;
                 bestMove = position;
             }
         }
