+++ src/main/java/application/players/Player.java
@@ -1,6 +1,9 @@
-package application;
+package application.players;
 
-interface Player {
+import application.game.Board;
+import application.game.COLOUR;
+
+public interface Player {
 
     Board playTurn(Board board);
