+++ src/main/java/application/utils/GsonFactory.java
@@ -1,4 +1,4 @@
-package utils;
+package application.utils;
 
 
 import com.google.gson.Gson;
